chore: [DHIS2-12142] useListAttributes
